# docs/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Cook Book'
  version: ''
paths:
...
  /api/users/subscriptions/:
    get:
      operationId: Мои подписки
      description: 'Возвращает пользователей, на которых подписан текущий пользователь. В выдачу добавляются рецепты.'
      parameters:
        - name: page
          required: false
          in: query
          description: Номер страницы.
          schema:
            type: integer
        - name: limit
          required: false
          in: query
          description: Количество объектов на странице.
          schema:
            type: integer
        - name: recipes_limit
          required: false
          in: query
          description: Количество объектов внутри поля recipes.
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 123
                    description: 'Общее количество объектов в базе'
                  next:
                    type: string
                    nullable: true
                    format: uri
                    example: http://cookbook.example.org/api/users/subscriptions/?page=4
                    description: 'Ссылка на следующую страницу'
                  previous:
                    type: string
                    nullable: true
                    format: uri
                    example: http://cookbook.example.org/api/users/subscriptions/?page=2
                    description: 'Ссылка на предыдущую страницу'
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserWithRecipes'
                    description: 'Список объектов текущей страницы'
          description: ''
        '401':
          $ref: '#/components/responses/AuthenticationError'
      tags:
        - Подписки
  /api/users/{id}/subscribe/:
    post:
      operationId: Подписаться на пользователя
      description: 'Доступно только авторизованным пользователям'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого пользователя."
          schema:
            type: string
        - name: recipes_limit
          required: false
          in: query
          description: Количество объектов внутри поля recipes.
          schema:
            type: integer
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithRecipes'
          description: 'Подписка успешно создана'
        '400':
          description: 'Ошибка подписки (Например, если уже подписан или при подписке на себя самого)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfMadeError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'
      tags:
        - Подписки
    delete:
      operationId: Отписаться от пользователя
      description: 'Доступно только авторизованным пользователям'
      security:
        - Token: [ ]
      parameters:
        - name: id
          in: path
          required: true
          description: "Уникальный идентификатор этого пользователя."
          schema:
            type: string
      responses:
        '204':
          description: 'Успешная отписка'
        '400':
          description: 'Ошибка отписки (Например, если не был подписан)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SelfMadeError'
        '401':
          $ref: '#/components/responses/AuthenticationError'
        '404':
          $ref: '#/components/responses/NotFound'

      tags:
        - Подписки
...
components:
  schemas:
...
    UserWithRecipes:
      description: 'Расширенный объект пользователя с рецептами'
      type: object
      properties:
        email:
          type: string
          format: email
          maxLength: 254
          description: "Адрес электронной почты"
        id:
          type: integer
          readOnly: true
        username:
          type: string
          description: "Уникальный юзернейм"
          pattern: ^[\w.@+-]+\z
          maxLength: 150
        first_name:
          type: string
          maxLength: 150
          description: "Имя"
          example: "Вася"
        last_name:
          type: string
          maxLength: 150
          description: "Фамилия"
          example: "Пупкин"
        is_subscribed:
          type: boolean
          readOnly: true
          description: "Подписан ли текущий пользователь на этого"
        recipes:
          type: array
          items:
            $ref: '#/components/schemas/RecipeMinified'
        recipes_count:
          type: integer
          description: 'Общее количество рецептов пользователя'
...
    RecipeMinified:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: 'Уникальный id'
        name:
          type: string
          maxLength: 200
          description: 'Название'
        image:
          description: 'Ссылка на картинку на сайте'
          example: 'http://cookbook.example.org/media/recipes/images/image.jpeg'
          type: string
          format: url
        cooking_time:
          description: 'Время приготовления (в минутах)'
          type: integer
          minimum: 1
...
    SelfMadeError:
      description: Ошибка
      type: object
      properties:
        errors:
          description: 'Описание ошибки'
          type: string
    AuthenticationError:
      description: Пользователь не авторизован
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Учетные данные не были предоставлены."
          type: string
...
    NotFound:
      description: Объект не найден
      type: object
      properties:
        detail:
          description: 'Описание ошибки'
          example: "Страница не найдена."
          type: string
  responses:
...
    AuthenticationError:
      description: Пользователь не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthenticationError'
...
    NotFound:
      description: Объект не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
  securitySchemes:
    Token:
      description: 'Авторизация по токену. <br>
      Все запросы от имени пользователя должны выполняться с заголовком "Authorization: Token TOKENVALUE"'
      type: http
      scheme: token